Adding a 'get_nth_value' function.
