+++ tests/FileOperations/test_file_operations.py
@@ -1,0 +1,122 @@
+import pytest
+from pytest_mock import MockerFixture
+
+import ResSimpy.FileOperations.file_operations as fo
+
+
+@pytest.mark.parametrize("line_contents, file_contents, expected_result", [
+    ("MYTESTTOKEN 3", "MYTESTTOKEN 3\n ANOTHER_TOKEN 8", '3'),
+    ("MYTESTTOKEN 123",
+     """GOTABLE 
+               SG         KRG        KROG        PCGO
+          0.000000     0.000000     1.000000     0.000000
+          0.085260     0.011210     0.681265     0.000000
+          ! Test comment in the middle of the table          
+          0.100598     0.014109     0.625140     0.000000
+          ANOTHERTOKEN 6
+        MYTESTTOKEN 123
+          TOKENCONTAINING_MYTESTTOKEN 8
+          FINALTOKEN 90
+
+          """,
+     '123'),
+    ("MYTESTTOKEN",
+     """GOTABLE 
+       SG         KRG        KROG        PCGO
+  0.000000     0.000000     1.000000     0.000000
+  0.085260     0.011210     0.681265     0.000000
+  ! Test comment in the middle of the table    
+C Another comment using 'C'        
+  0.100598     0.014109     0.625140     0.000000
+  ANOTHERTOKEN 6
+MYTESTTOKEN 
+7
+  TOKENCONTAINING_MYTESTTOKEN 8
+  FINALTOKEN 90
+
+  """,
+     '7'),
+    ("MYTESTTOKEN",
+     '''MYTESTTOKEN
+     C Comment line
+     token_value''',
+     'token_value'),
+    ("not a comment C MYTESTTOKEN",
+     '''not a comment C MYTESTTOKEN
+     C Comment line
+     C
+     Ctoken_value''',
+     'Ctoken_value'),
+    ("MYTESTTOKEN",
+     '''MYTESTTOKEN
+     C Comment line
+     "token value"''',
+     'token value'),
+
+], ids=['basic case', 'multiple lines', 'value on next line', 'Comment character C', 'complex C comment'
+    , 'get value in double quotes'])
+def test_get_token_value(mocker: MockerFixture, line_contents, file_contents, expected_result):
+    # Arrange
+    dummy_file_as_list = [y for y in (x.strip() for x in file_contents.splitlines()) if y]
+    open_mock = mocker.mock_open(read_data=file_contents)
+    mocker.patch("builtins.open", open_mock)
+
+    # Act
+    result = fo.get_token_value(token='MYTESTTOKEN', token_line=line_contents,
+                                file_list=dummy_file_as_list)
+
+    # Assert
+    assert result == expected_result
+
+
+@pytest.mark.parametrize("line, number_tokens, expected_result", [
+    ('EQUIL METHOD 1 /path/equil.dat', 4, ['EQUIL', 'METHOD', '1', '/path/equil.dat']),
+    ('EQUIL METHOD 1 /path/equil.dat ! comment', 4, ['EQUIL', 'METHOD', '1', '/path/equil.dat']),
+    ('EQUIL NorPT METHOD 1 /path/equil.dat TOKEN TOKEN', 6,
+     ['EQUIL', 'METHOD', '1', '/path/equil.dat', 'TOKEN', 'TOKEN']),
+    ('EQUIL METHOD 1 \n /path/equil.dat', 4, ['EQUIL', 'METHOD', '1', '/path/equil.dat']),
+    ('EQUIL METHOD !comment\n \t 1 ', 3, ['EQUIL', 'METHOD', '1']),
+    ('EQUIL\n NORPT METHOD\n1\n/path/equil.dat', 4, ['EQUIL', 'METHOD', '1', '/path/equil.dat']),
+    ('EQUIL METHOD 1 /path/equil.dat', 2, ['EQUIL', 'METHOD']),
+    ('\n \n \n EQUIL METHOD 1 /path/equil.dat', 2, ['EQUIL', 'METHOD']),
+], ids=["basic", "more tokens", "get more tokens", "new line", "newline comment", "lots of newlines",
+        "more text than declared tokens", "starting with new lines"])
+def test_get_multiple_sequential_values(line, number_tokens, expected_result):
+    # Arrange
+    list_of_strings = line.splitlines()
+    # Act
+    result = fo.get_multiple_expected_sequential_values(list_of_strings, number_tokens, ['NORPT'])
+    # Assert
+    assert result == expected_result
+
+
+def test_get_multiple_sequential_values_fail_case():
+    # Arrange
+    line = 'EQUIL METHOD \n \n \n'
+    number_tokens = 3
+    expected_error_str = ('Too many values requested from the list of strings passed, instead found: 2 values, '
+                          'out of the requested 3')
+    # Act + Assert
+    with pytest.raises(ValueError) as ve:
+        value = fo.get_multiple_expected_sequential_values(line.splitlines(), number_tokens, [])
+    result_error_msg = str(ve.value)
+    assert result_error_msg == expected_error_str
+
+
+@pytest.mark.parametrize("list_of_strings, value_number_to_get, expected_result", [
+    (["a"], 1, 'a'),
+    ([" 1 2 \t 3 4"], 3, '3'),
+    ([" 1 2 \t 3 4"], 1, '1'),
+    ([" A 2 \t C 4"], 3, 'C'),
+    ([" 1 2 \t 3"], 4, None),
+    ([" 1 ", "2 \t 3 ", "4"], 3, '3'),
+    ([" 1 ", "2 \t 3 ", "4"], 4, '4'),
+    ([" Value1 ", "Value2 \t", "   ", " Value3 ", "Value4"], 3, 'Value3'),
+    ([], 1, None),
+])
+def test_get_nth_value(list_of_strings, value_number_to_get, expected_result):
+    # Act
+    result = fo.get_nth_value(list_of_strings=list_of_strings, value_number=value_number_to_get, ignore_values=[])
+
+    # Assert
+    assert result == expected_result
